+++ src/classes/index.ts
@@ -1,3 +1,4 @@
+export * from "./ApplicationCommandMixin";
 export * from "./DefaultPermissionResolver";
 export * from "./SimpleCommandMessage";
 export * from "./errors";